remove mkl baxed CSR creation

+++ cpp/knng/knng_handler.hpp
@@ -193,7 +193,7 @@ public:
     (*output_knng).resize(total_nn_size);
 
     #pragma omp parallel for schedule (static)
-    for(size_t i = 0; i < final_nn_map.size(); ++i){
+    for(uint64_t i = 0; i < final_nn_map.size(); ++i){
       auto it = final_nn_map.begin() + i;
       vector<EdgeNode<INDEX_TYPE,VALUE_TYPE>> edge_node_list = (*it).second;
       auto pre_index = i*(nn-1);
